+++ rlap/cc/factorizers.h
@@ -13,10 +13,10 @@ class Factorizer{
   public:
     // Abstract base class for all factorizers
     virtual void compute() = 0;
-    virtual Eigen::SparseMatrix<float> getLaplacian() = 0;
-    virtual Eigen::MatrixXf getReconstructedLaplacian() = 0;
+    virtual Eigen::SparseMatrix<double> getLaplacian() = 0;
+    virtual Eigen::MatrixXd getReconstructedLaplacian() = 0;
     // get the laplacian of the adjacency matrix
-    Eigen::SparseMatrix<float>* computeLaplacian(Eigen::SparseMatrix<float>* Adj);
+    Eigen::SparseMatrix<double>* computeLaplacian(Eigen::SparseMatrix<double>* Adj);
     virtual ~Factorizer() = default;
 };
 
@@ -25,71 +25,71 @@ class Factorizer{
 class EigenCholeskyLLT: public Factorizer{
   public:
     // Eigen implementation of the L * L^t cholesky factorization
-    EigenCholeskyLLT(Eigen::SparseMatrix<float>* Adj);
+    EigenCholeskyLLT(Eigen::SparseMatrix<double>* Adj);
     ~EigenCholeskyLLT(){};
     // compute the L * L^T based cholesky factorization
     void compute() override;
     // retrive the computed laplacian
-    Eigen::SparseMatrix<float> getLaplacian() override;
+    Eigen::SparseMatrix<double> getLaplacian() override;
     // retrieve the reconstructed matrix
-    Eigen::MatrixXf getReconstructedLaplacian() override;
+    Eigen::MatrixXd getReconstructedLaplacian() override;
   private:
-    Eigen::SparseMatrix<float>* _L;
-    Eigen::LLT<Eigen::MatrixXf>* _llt;
+    Eigen::SparseMatrix<double>* _L;
+    Eigen::LLT<Eigen::MatrixXd>* _llt;
 };
 
 class EigenCholeskyLDLT: public Factorizer{
   public:
     // Eigen implementation of the L * D * L^t cholesky factorization
-    EigenCholeskyLDLT(Eigen::SparseMatrix<float>* Adj);
+    EigenCholeskyLDLT(Eigen::SparseMatrix<double>* Adj);
     ~EigenCholeskyLDLT(){};
     // compute the L * D * L^T based cholesky factorization
     void compute() override;
     // retrive the computed laplacian
-    Eigen::SparseMatrix<float> getLaplacian() override;
+    Eigen::SparseMatrix<double> getLaplacian() override;
     // retrieve the reconstructed matrix
-    Eigen::MatrixXf getReconstructedLaplacian() override;
+    Eigen::MatrixXd getReconstructedLaplacian() override;
   private:
-    Eigen::SparseMatrix<float>* _L;
-    Eigen::LDLT<Eigen::MatrixXf>* _ldlt;
+    Eigen::SparseMatrix<double>* _L;
+    Eigen::LDLT<Eigen::MatrixXd>* _ldlt;
 };
 
 class ClassicCholesky: public Factorizer{
   public:
     // Standard cholesky factorization from the ground up
-    ClassicCholesky(Eigen::SparseMatrix<float>* Adj);
+    ClassicCholesky(Eigen::SparseMatrix<double>* Adj);
     ~ClassicCholesky(){};
     // compute the L * L^T based cholesky factorization
     void compute() override;
     // retrive the computed laplacian
-    Eigen::SparseMatrix<float> getLaplacian() override;
+    Eigen::SparseMatrix<double> getLaplacian() override;
     // retrieve the reconstructed matrix
-    Eigen::MatrixXf getReconstructedLaplacian() override;
+    Eigen::MatrixXd getReconstructedLaplacian() override;
     // retrive the factorized lower triangular matrix
-    Eigen::SparseMatrix<float> getLower();
+    Eigen::SparseMatrix<double> getLower();
   private:
-    Eigen::SparseMatrix<float>* _L;
-    Eigen::SparseMatrix<float>* _G;
+    Eigen::SparseMatrix<double>* _L;
+    Eigen::SparseMatrix<double>* _G;
 };
 
 class NaiveApproximateCholesky: public Factorizer{
   public:
     // approximate cholesky factorization from the ground up
-    NaiveApproximateCholesky(Eigen::SparseMatrix<float>* Adj);
+    NaiveApproximateCholesky(Eigen::SparseMatrix<double>* Adj);
     ~NaiveApproximateCholesky(){};
     // compute the approximate L * L^T based cholesky factorization.
     // Please note that this method is just for exploration and understanding
     // purposes. It doesn't scale well practically.
     void compute() override;
     // retrive the computed laplacian
-    Eigen::SparseMatrix<float> getLaplacian() override;
+    Eigen::SparseMatrix<double> getLaplacian() override;
     // retrieve the reconstructed matrix
-    Eigen::MatrixXf getReconstructedLaplacian() override;
+    Eigen::MatrixXd getReconstructedLaplacian() override;
   private:
-    Eigen::SparseMatrix<float>* getStar(Eigen::SparseMatrix<float>* L, int i);
-    Eigen::SparseMatrix<float>* _A;
-    Eigen::SparseMatrix<float>* _L;
-    Eigen::SparseMatrix<float>* _G;
+    Eigen::SparseMatrix<double>* getStar(Eigen::SparseMatrix<double>* L, int i);
+    Eigen::SparseMatrix<double>* _A;
+    Eigen::SparseMatrix<double>* _L;
+    Eigen::SparseMatrix<double>* _G;
 };
 
 // end of reference/benchmark factorizers
@@ -100,8 +100,8 @@ class NaiveApproximateCholesky: public Factorizer{
 class ApproximateCholesky: public Factorizer{
   public:
     // approximate cholesky factorization from the ground up
-    ApproximateCholesky(Eigen::SparseMatrix<float>* Adj, std::string pre = "degree");
-    ApproximateCholesky(Eigen::SparseMatrix<float> Adj, std::string pre = "degree");
+    ApproximateCholesky(Eigen::SparseMatrix<double>* Adj, std::string pre = "degree");
+    ApproximateCholesky(Eigen::SparseMatrix<double> Adj, std::string pre = "degree");
     // A constructor which enables the user to just pass the
     // tsv/csv file for the adjacency matrix of the graph.
     // This reduces the overhead of copying scipy csc matrices
@@ -110,20 +110,20 @@ class ApproximateCholesky: public Factorizer{
     ~ApproximateCholesky(){};
     // retrieve the adjancency matrix. helpful if factorizer was
     // initialized with a filename and dimension
-    Eigen::SparseMatrix<float> getAdjacencyMatrix();
+    Eigen::SparseMatrix<double> getAdjacencyMatrix();
     // compute the pre-conditioning info for solvers
     void compute() override;
     // retrive the computed laplacian
-    Eigen::SparseMatrix<float> getLaplacian() override;
+    Eigen::SparseMatrix<double> getLaplacian() override;
     // retrieve the reconstructed matrix
-    Eigen::MatrixXf getReconstructedLaplacian() override;
+    Eigen::MatrixXd getReconstructedLaplacian() override;
     // retrieve the pre-conditioner
     LDLi* getPreconditioner();
     // solve for the unknowns
-    Eigen::VectorXf solve(Eigen::VectorXf b);
+    Eigen::VectorXd solve(Eigen::VectorXd b);
   private:
-    Eigen::SparseMatrix<float>* _A;
-    Eigen::SparseMatrix<float>* _L;
+    Eigen::SparseMatrix<double>* _A;
+    Eigen::SparseMatrix<double>* _L;
     LDLi* _ldli;
     std::string _pre_str;
 };
