+++ rlap/cc/preconditioner.h
@@ -13,39 +13,39 @@ class Preconditioner{
 
 class OrderedPreconditioner : public Preconditioner{
   public:
-    OrderedPreconditioner(Eigen::SparseMatrix<float>* A);
+    OrderedPreconditioner(Eigen::SparseMatrix<double>* A);
     ~OrderedPreconditioner(){};
     LDLi* getLDLi();
   private:
     OrderedMatrix* getOrderedMatrix();
-    float getColumnLength(OrderedMatrix* ordmat, int i, std::vector<ColumnElement>* colspace);
+    double getColumnLength(OrderedMatrix* ordmat, int i, std::vector<ColumnElement>* colspace);
     void printColumn(OrderedMatrix* ordmat, int i);
     void printColspace(std::vector<ColumnElement>* colspace);
     void printMatrixDetails(OrderedMatrix* a);
-    float compressColumn(std::vector<ColumnElement>* colspace, float len);
-    Eigen::SparseMatrix<float>* _A;
+    double compressColumn(std::vector<ColumnElement>* colspace, double len);
+    Eigen::SparseMatrix<double>* _A;
     OrderedMatrix* _ordmat;
 };
 
 class PriorityPreconditioner : public Preconditioner{
   public:
-    PriorityPreconditioner(Eigen::SparseMatrix<float>* A);
+    PriorityPreconditioner(Eigen::SparseMatrix<double>* A);
     ~PriorityPreconditioner(){};
     LDLi* getLDLi();
   private:
     PriorityMatrix* getPriorityMatrix();
     void printColumn(PriorityMatrix* pmat, int i);
-    DegreePQ* getDegreePQ(std::vector<float> a);
-    float DegreePQPop(DegreePQ* pq);
-    void DegreePQMove(DegreePQ* pq, int i, float newkey, int oldlist, int newlist);
+    DegreePQ* getDegreePQ(std::vector<double> a);
+    double DegreePQPop(DegreePQ* pq);
+    void DegreePQMove(DegreePQ* pq, int i, double newkey, int oldlist, int newlist);
     void DegreePQDec(DegreePQ* pq, int i);
     void DegreePQInc(DegreePQ* pq, int i);
-    std::vector<float> getFlipIndices(Eigen::SparseMatrix<float>* M);
-    float getColumnLength(PriorityMatrix* pmat, int i, std::vector<PriorityElement*>* colspace);
-    float compressColumn(PriorityMatrix* pmat, std::vector<PriorityElement*>* colspace, float len, DegreePQ* pq);
-    void printFlipIndices(std::vector<float> fi);
+    std::vector<double> getFlipIndices(Eigen::SparseMatrix<double>* M);
+    double getColumnLength(PriorityMatrix* pmat, int i, std::vector<PriorityElement*>* colspace);
+    double compressColumn(PriorityMatrix* pmat, std::vector<PriorityElement*>* colspace, double len, DegreePQ* pq);
+    void printFlipIndices(std::vector<double> fi);
 
-    Eigen::SparseMatrix<float>* _A;
+    Eigen::SparseMatrix<double>* _A;
     PriorityMatrix* _pmat;
 };
